requirements.txt
discord.py
cryptography
pynacl
aiohttp
aiohttp_session
